+++ app/src/main/AndroidManifest.xml
@@ -20,7 +20,16 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name="com.dhakaregency.KOT"></activity>
+        <activity android:name="com.dhakaregency.KOT" />
+        <activity
+            android:name="com.dhakaregency.Modules"
+            android:label="@string/title_activity_modules"
+            android:theme="@style/AppTheme.NoActionBar" />
+        <!-- ATTENTION: This was auto-generated to add Google Play services to your project for
+             App Indexing.  See https://g.co/AppIndexing/AndroidStudio for more information. -->
+        <meta-data
+            android:name="com.google.android.gms.version"
+            android:value="@integer/google_play_services_version" />
     </application>
 
 </manifest>
