+++ app/src/main/java/com/dhakaregency/OutletsEntity.java
@@ -1,0 +1,30 @@
+package com.dhakaregency;
+
+/**
+ * Created by Administrator on 26/02/2016.
+ */
+public class OutletsEntity {
+    private String id;
+    private String name;
+
+    public OutletsEntity()
+    {
+
+    }
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+}
