+++ app/src/main/java/com/dhakaregency/SignIn.java
@@ -1,23 +1,28 @@
 package com.dhakaregency;
 
-import android.app.ProgressDialog;
+import android.os.AsyncTask;
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-import android.util.Log;
 import android.view.View;
 
 import com.dhakaregency.quickkot.R;
 import android.content.Intent;
-import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.TextView;
-import android.widget.Toast;
 
+import org.json.JSONException;
+import org.json.JSONObject;
+import org.json.JSONStringer;
+
+import java.io.BufferedReader;
+import java.io.InputStreamReader;
+import java.io.OutputStreamWriter;
+import java.net.HttpURLConnection;
+import java.net.MalformedURLException;
+import java.net.URL;
 import java.util.ArrayList;
+
+import javax.net.ssl.HttpsURLConnection;
 
 public class SignIn extends AppCompatActivity {
 
@@ -28,16 +33,15 @@ public class SignIn extends AppCompatActivity {
     EditText _passwordText;
     Button _loginButton;
 
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
 
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_sign_in);
+
         _useridText=(EditText)findViewById(R.id.input_userid);
         _passwordText=(EditText)findViewById(R.id.input_password);
         _loginButton=(Button)findViewById(R.id.btn_login);
-
         _loginButton.setOnClickListener(new View.OnClickListener() {
 
             @Override
@@ -47,7 +51,7 @@ public class SignIn extends AppCompatActivity {
         });
     }
     public void login() {
-        Log.d(TAG, "Login");
+
 
         if (!validate()) {
             onLoginFailed();
@@ -56,17 +60,8 @@ public class SignIn extends AppCompatActivity {
 
         _loginButton.setEnabled(false);
 
-        final ProgressDialog progressDialog = new ProgressDialog(this,
-                ProgressDialog.STYLE_HORIZONTAL);
-        progressDialog.setIndeterminate(true);
-        progressDialog.setMessage("Authenticating...");
-        progressDialog.show();
-
         String user_id = _useridText.getText().toString();
         String password = _passwordText.getText().toString();
-
-        // TODO: Implementing Web Service Authentication Here
-// calling User Authentication Service Class
 
         ArrayList<String> passing = new ArrayList<String>();
         passing.add(user_id);
@@ -102,31 +97,120 @@ public class SignIn extends AppCompatActivity {
     }
 
     public void onLoginSuccess() {
-        _loginButton.setEnabled(true);
-        finish();
+        Intent intent = new Intent(this,Modules.class );
+        startActivity(intent);
+
     }
 
     public void onLoginFailed() {
-        Toast.makeText(getBaseContext(), "Login failed", Toast.LENGTH_LONG).show();
-
         _loginButton.setEnabled(true);
     }
 
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        if (requestCode == REQUEST_SIGNUP) {
-            if (resultCode == RESULT_OK) {
 
-                // TODO: Implement successful signup logic here
-                // By default we just finish the Activity and log them in automatically
-                this.finish();
+
+    public class UserAuthCheck extends AsyncTask<ArrayList<String>, Void, Integer> {
+
+        @Override
+        protected void onPreExecute() {
+            super.onPreExecute();
+
+        }
+
+        @Override
+        protected void onPostExecute(Integer i) {
+            super.onPostExecute(i);
+
+            if (i==1)
+            {
+            onLoginSuccess();
+                  }
+            else
+            {
+                onLoginFailed();
             }
         }
-    }
 
-    @Override
-    public void onBackPressed() {
-        // disable going back to the MainActivity
-        moveTaskToBack(true);
+        @Override
+        protected Integer  doInBackground(ArrayList<String>... params) {
+
+            String str = "http://192.168.99.12:8080/AuthService.svc/GetAuth";
+            String response="";
+            URL url=null;
+            try {
+                url = new URL(str);
+            } catch (MalformedURLException e) {
+                e.printStackTrace();
+            }
+            try {
+
+                HttpURLConnection conn=null;
+                try {
+                    conn = (HttpURLConnection) url.openConnection();
+                }
+                catch (Exception ex)
+                {
+                    ex.printStackTrace();
+                }
+                ArrayList<String> passed = params[0];
+
+                conn.setReadTimeout(15000);
+                conn.setConnectTimeout(15000);
+                conn.setRequestMethod("POST");
+                conn.setDoInput(true);
+                conn.setDoOutput(true);
+
+                conn.setRequestProperty("Content-Type", "application/json");
+
+
+                JSONObject jsonObject = new JSONObject();
+                // Build JSON string
+                JSONStringer userJson = new JSONStringer()
+                        .object()
+                        .key("user")
+                        .object()
+                        .key("userid").value(passed.get(0).toString())
+                        .key("password").value(passed.get(1).toString())
+                        .endObject()
+                        .endObject();
+
+                //byte[] outputBytes = jsonParam.toString().getBytes("UTF-8");
+                OutputStreamWriter outputStreamWriter = new OutputStreamWriter(conn.getOutputStream());
+                outputStreamWriter.write(userJson.toString());
+                outputStreamWriter.close();
+
+                int responseCode=conn.getResponseCode();
+
+                if (responseCode == HttpsURLConnection.HTTP_OK) {
+                    String line;
+                    BufferedReader br=new BufferedReader(new InputStreamReader(conn.getInputStream()));
+                    while ((line=br.readLine()) != null) {
+                        response+=line;
+                    }
+                }
+                else {
+                    response="";
+
+                }
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+            Integer result=0;
+            JSONObject jObject=null;
+            if (!response.isEmpty()) {
+                try {
+                    jObject = new JSONObject(response);
+                } catch (JSONException e) {
+                    e.printStackTrace();
+                }
+                try {
+                    result= Integer.parseInt(jObject.getString("GetAuthResult"));
+                } catch (JSONException e) {
+                    e.printStackTrace();
+                }
+                ;
+            }
+            return result;
+        }
+
     }
 }
