+++ app/src/main/res/layout/content_sign_in.xml
@@ -29,6 +29,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:inputType="textPersonName"
+                android:text="Admin"
                 android:hint="User Id" />
         </android.support.design.widget.TextInputLayout>
 
@@ -42,6 +43,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:inputType="textPassword"
+                android:text="laura"
                 android:hint="Password"/>
         </android.support.design.widget.TextInputLayout>
