+++ app/src/main/res/values/strings.xml
@@ -4,6 +4,8 @@
     <string name="title_activity_login">Login</string>
     <string name="title_activity_sign_in">Quick KOT</string>
     <string name="user_id">User Id</string>
+    <string name="title_activity_modules">modules</string>
+    <string name="title_activity_appsettings">appsettings</string>
 
 
 </resources>
